Simplify group join route to /group/[token]/join with OTP verification
Move join route from /join/[token] to /group/[token]/join with a multi-step
flow: contact form → SMS OTP verification → optional group password → join.

- Extract shared OTPVerifyForm component from onboarding verify page
- Refactor onboarding/verify to use shared OTPVerifyForm (no duplication)
- Create new /group/[token]/join page with form → OTP → password steps
- Old /join/[token] route redirects to new URL for backwards compatibility
- Update all share URL references across group-detail, group-settings,
  group-list, and robots.ts
