+++ src/app/join/[token]/page.tsx
@@ -1,405 +1,10 @@
-'use client'
+import { redirect } from 'next/navigation'
 
-import { use, useEffect, useState } from 'react'
-import { useForm, Controller } from 'react-hook-form'
-import { zodResolver } from '@hookform/resolvers/zod'
-import { useQuery, useMutation } from '@tanstack/react-query'
-import type { User } from '@supabase/supabase-js'
-import { supabase } from '@/lib/supabase'
-import type { ContactGroup, Database } from '@/types'
-import { getGroupByToken, joinContactGroup, joinContactGroupAnonymous, logShareLinkView } from '@/lib/database'
-import { Button } from '@/components/ui/button'
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card'
-import { Input } from '@/components/ui/input'
-import { PhoneInput } from '@/components/ui/phone-input'
-import { Checkbox } from '@/components/ui/checkbox'
-import { ArrowLeft, Users, CheckCircle, Lock } from 'lucide-react'
-import Link from 'next/link'
-import { toast } from 'sonner'
-import { contactFormSchema, type ContactFormData } from '@/lib/validations'
-
-// Force dynamic rendering
-export const dynamic = 'force-dynamic'
-
-interface JoinPageProps {
-  params: Promise<{
-    token: string
-  }>
+interface JoinRedirectProps {
+  params: Promise<{ token: string }>
 }
 
-export default function JoinPage({ params }: JoinPageProps) {
-  const resolvedParams = use(params)
-  const token = resolvedParams.token
-  type ProfileRow = Database['public']['Tables']['profiles']['Row']
-
-  const [hasJoined, setHasJoined] = useState(false)
-  const [user, setUser] = useState<User | null>(null)
-
-  const form = useForm<ContactFormData>({
-    resolver: zodResolver(contactFormSchema),
-    defaultValues: {
-      first_name: '',
-      last_name: '',
-      email: '',
-      phone: '',
-      notifications_enabled: false,
-      group_password: '',
-    },
-  })
-
-  // Check if user is logged in (optional, don't block on this)
-  useEffect(() => {
-    const checkUser = async () => {
-      try {
-        const { data: { user } } = await supabase.auth.getUser()
-        if (user) {
-          setUser(user)
-          // Try to get profile, but don't block if it fails
-          const { data: profile } = await supabase
-            .from('profiles')
-            .select('first_name,last_name,email,phone,sms_notifications_enabled')
-            .eq('id', user.id)
-            .single<ProfileRow>()
-
-          if (profile) {
-            form.reset({
-              first_name: profile.first_name || '',
-              last_name: profile.last_name || '',
-              email: profile.email || user.email || '',
-              phone: profile.phone || '',
-              notifications_enabled: profile.sms_notifications_enabled || false,
-              group_password: '',
-            })
-          }
-        }
-      } catch (error) {
-        console.log('No user logged in or profile fetch failed:', error)
-        // Continue without auth - this is fine for anonymous users
-      }
-    }
-
-    checkUser()
-  }, [form])
-
-  // Fetch group details
-  const { data: group, isLoading: groupLoading, error: groupError } = useQuery<ContactGroup | null>({
-    queryKey: ['group-by-token', token],
-    queryFn: async () => {
-      const result = await getGroupByToken(token)
-      if (result.error) throw new Error(result.error)
-      return result.data
-    },
-  })
-
-  // Log share link view — fire-and-forget, non-blocking
-  useEffect(() => {
-    if (!group?.id) return
-    const params = new URLSearchParams(window.location.search)
-    logShareLinkView(token, {
-      referrer: document.referrer || undefined,
-      utm_source: params.get('utm_source') || undefined,
-      utm_medium: params.get('utm_medium') || undefined,
-      utm_campaign: params.get('utm_campaign') || undefined,
-    })
-  }, [group?.id, token])
-
-  // Join group mutation
-  const requiresPassword = group?.access_type === 'password'
-
-  const joinMutation = useMutation({
-    mutationFn: async () => {
-      const values = form.getValues()
-      const password = requiresPassword ? values.group_password : undefined
-      if (user) {
-        // Authenticated user
-        const result = await joinContactGroup(token, values.notifications_enabled, password)
-        if (result.error) throw new Error(result.error)
-        return result.data
-      } else {
-        // Anonymous user
-        const result = await joinContactGroupAnonymous(
-          token,
-          values.first_name,
-          values.last_name,
-          values.email,
-          values.phone,
-          values.notifications_enabled,
-          password
-        )
-        if (result.error) throw new Error(result.error)
-        return result.data
-      }
-    },
-    onSuccess: () => {
-      setHasJoined(true)
-      toast.success('Successfully joined the group!')
-    },
-    onError: (error: unknown) => {
-      const errorMessage = error instanceof Error ? error.message : 'Failed to join group'
-      toast.error(errorMessage)
-    },
-  })
-
-  const onSubmit = form.handleSubmit((values) => {
-    if (requiresPassword && !values.group_password?.trim()) {
-      toast.error('A group password is required to join this group.')
-      return
-    }
-    joinMutation.mutate()
-  })
-
-  if (groupLoading) {
-    return (
-      <div className="min-h-screen bg-background flex items-center justify-center">
-        <div className="animate-spin rounded-full h-8 w-8 border-b-2 border-primary"></div>
-      </div>
-    )
-  }
-
-  if (groupError || !group) {
-    return (
-      <div className="min-h-screen bg-background flex items-center justify-center p-4">
-        <Card className="w-full max-w-md">
-          <CardHeader>
-            <CardTitle className="text-center text-destructive">Group Not Found</CardTitle>
-          </CardHeader>
-          <CardContent className="text-center">
-            <p className="text-muted-foreground mb-4">
-              The group link is invalid or the group no longer exists.
-            </p>
-            <Link href="/">
-              <Button variant="outline">
-                <ArrowLeft className="h-4 w-4 mr-2" />
-                Go Home
-              </Button>
-            </Link>
-          </CardContent>
-        </Card>
-      </div>
-    )
-  }
-
-  if (group.is_closed) {
-    return (
-      <div className="min-h-screen bg-background flex items-center justify-center p-4">
-        <Card className="w-full max-w-md">
-          <CardHeader>
-            <CardTitle className="text-center">Group Closed</CardTitle>
-          </CardHeader>
-          <CardContent className="text-center">
-            <p className="text-muted-foreground mb-4">
-              This group is no longer accepting new members.
-            </p>
-            <Link href="/">
-              <Button variant="outline">
-                <ArrowLeft className="h-4 w-4 mr-2" />
-                Go Home
-              </Button>
-            </Link>
-          </CardContent>
-        </Card>
-      </div>
-    )
-  }
-
-  if (group.archived_at) {
-    return (
-      <div className="min-h-screen bg-background flex items-center justify-center p-4">
-        <Card className="w-full max-w-md">
-          <CardHeader>
-            <CardTitle className="text-center">Group Not Available</CardTitle>
-          </CardHeader>
-          <CardContent className="text-center">
-            <p className="text-muted-foreground mb-4">
-              This group is no longer available.
-            </p>
-            <Link href="/">
-              <Button variant="outline">
-                <ArrowLeft className="h-4 w-4 mr-2" />
-                Go Home
-              </Button>
-            </Link>
-          </CardContent>
-        </Card>
-      </div>
-    )
-  }
-
-  if (hasJoined) {
-    return (
-      <div className="min-h-screen bg-background flex items-center justify-center p-4">
-        <Card className="w-full max-w-md">
-          <CardHeader>
-            <CardTitle className="text-center flex items-center justify-center gap-2">
-              <CheckCircle className="h-6 w-6 text-[#065F46]" />
-              Successfully Joined!
-            </CardTitle>
-          </CardHeader>
-          <CardContent className="text-center">
-            <p className="text-muted-foreground mb-4">
-              You&apos;ve been added to <strong>{group.name}</strong>. Other members can now see your contact information.
-            </p>
-            {user ? (
-              <Link href="/dashboard">
-                <Button>Go to Dashboard</Button>
-              </Link>
-            ) : (
-              <div className="space-y-2">
-                <Link href="/auth?mode=signup">
-                  <Button className="w-full">Create Account</Button>
-                </Link>
-                <Link href="/">
-                  <Button variant="outline" className="w-full">Go Home</Button>
-                </Link>
-              </div>
-            )}
-          </CardContent>
-        </Card>
-      </div>
-    )
-  }
-
-  return (
-    <div className="min-h-screen bg-background flex items-center justify-center p-4">
-      <Card className="w-full max-w-md">
-        <CardHeader>
-          <CardTitle className="flex items-center gap-2">
-            <Users className="h-5 w-5" />
-            Join {group.name}
-          </CardTitle>
-          <CardDescription>
-            {group.description || 'Share your contact information with other group members.'}
-          </CardDescription>
-          {group.owner && (
-            <p className="text-sm text-muted-foreground">
-              Organized by {group.owner.first_name} {group.owner.last_name}
-            </p>
-          )}
-        </CardHeader>
-
-        <CardContent>
-          <form onSubmit={onSubmit} className="space-y-4">
-            <div className="grid grid-cols-2 gap-4">
-              <div className="space-y-2">
-                <label htmlFor="first_name" className="text-sm font-medium">
-                  First Name *
-                </label>
-                <Input
-                  id="first_name"
-                  type="text"
-                  placeholder="Enter your first name"
-                  {...form.register('first_name')}
-                />
-                {form.formState.errors.first_name && (
-                  <p className="text-sm text-destructive">{form.formState.errors.first_name.message}</p>
-                )}
-              </div>
-
-              <div className="space-y-2">
-                <label htmlFor="last_name" className="text-sm font-medium">
-                  Last Name *
-                </label>
-                <Input
-                  id="last_name"
-                  type="text"
-                  placeholder="Enter your last name"
-                  {...form.register('last_name')}
-                />
-                {form.formState.errors.last_name && (
-                  <p className="text-sm text-destructive">{form.formState.errors.last_name.message}</p>
-                )}
-              </div>
-            </div>
-
-            <div className="space-y-2">
-              <label htmlFor="email" className="text-sm font-medium">
-                Email *
-              </label>
-              <Input
-                id="email"
-                type="email"
-                placeholder="Enter your email"
-                {...form.register('email')}
-              />
-              {form.formState.errors.email && (
-                <p className="text-sm text-destructive">{form.formState.errors.email.message}</p>
-              )}
-            </div>
-
-            <div className="space-y-2">
-              <label htmlFor="phone" className="text-sm font-medium">
-                Phone Number *
-              </label>
-              <Controller
-                name="phone"
-                control={form.control}
-                render={({ field }) => (
-                  <PhoneInput
-                    {...field}
-                    id="phone"
-                    className={form.formState.errors.phone ? '[&_input]:border-red-500' : ''}
-                  />
-                )}
-              />
-              {form.formState.errors.phone && (
-                <p className="text-sm text-destructive">{form.formState.errors.phone.message}</p>
-              )}
-            </div>
-
-            {group?.access_type === 'password' && (
-              <div className="space-y-2">
-                <label htmlFor="group_password" className="text-sm font-medium flex items-center gap-2">
-                  <Lock className="h-4 w-4" />
-                  Group Password
-                </label>
-                <Input
-                  id="group_password"
-                  type="password"
-                  placeholder="Enter the group password"
-                  {...form.register('group_password')}
-                />
-                <p className="text-xs text-muted-foreground">
-                  A password is required before you can join this group.
-                </p>
-              </div>
-            )}
-
-            <div className="flex items-center space-x-2">
-              <Controller
-                name="notifications_enabled"
-                control={form.control}
-                render={({ field }) => (
-                  <Checkbox
-                    id="notifications"
-                    checked={field.value}
-                    onCheckedChange={(checked) => field.onChange(!!checked)}
-                  />
-                )}
-              />
-              <label htmlFor="notifications" className="text-sm">
-                Notify me when members join or leave this group
-              </label>
-            </div>
-
-            <Button
-              type="submit"
-              className="w-full"
-              disabled={joinMutation.isPending}
-            >
-              {joinMutation.isPending ? 'Joining...' : 'Join Group'}
-            </Button>
-          </form>
-
-          <div className="mt-4 text-center">
-            <Link href="/">
-              <Button variant="outline" size="sm">
-                <ArrowLeft className="h-4 w-4 mr-2" />
-                Back to Home
-              </Button>
-            </Link>
-          </div>
-        </CardContent>
-      </Card>
-    </div>
-  )
+export default async function JoinRedirect({ params }: JoinRedirectProps) {
+  const { token } = await params
+  redirect(`/group/${token}/join`)
 }
